feat(example): created RTNEventService to replace async event handling in other services
